third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/CustomElementRegistry-initialize-expected.txt

This is a testharness.js-based test.
[FAIL] initialize throws when the registry scoped is false and the root is the document.
  assert_throws_dom: function "() => window.customElements.initialize(document)" threw object "TypeError: window.customElements.initialize is not a function" that is not a DOMException NotSupportedError: property "code" is equal to undefined, expected 9
[FAIL] initialize throws when the registry scoped is false and the root document already has another registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] initialize is a function on both global and scoped CustomElementRegistry
  assert_equals: expected "function" but got "undefined"
[FAIL] initialize sets element.customElementRegistry to the global registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize does not set the registry of nested shadow tree to the global registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize sets element.customElementRegistry to a scoped registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize does not set descendants whose customElementRegistry already uses a different registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize does not set the registry of nested shadow tree to a scoped registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize sets element.customElementRegistry permantently
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize is no-op on a subtree with a non-null registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] initialize works on Document
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize works on DocumentFragment
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize sets registry on shadow root descendants with no registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
Harness: the test ran to completion.
